+++ app.rb
@@ -5,9 +5,12 @@ require_relative 'rental'
 require_relative 'student'
 require_relative 'teacher'
 require './modules/data_loader'
+require './modules/data_saver'
 
 class App
   include DataLoader
+  include DataSaver
+
   def initialize
     @people = [] # Array to store people (teachers and students)
     @books = [] # Array to store books
@@ -129,40 +132,5 @@ class App
     load_people
     load_books
     load_rentals
-  end
-
-  private
-
-  def save_people
-    people_data = []
-
-    @people.map do |person|
-      if person.instance_of?(Student)
-        people_data.push({ name: person.name, age: person.age,
-                           parents_permission: person.parents_permission,
-                           classroom: person.classroom, type: 'Student' })
-      elsif person.instance_of?(Teacher)
-        people_data.push({ name: person.name, age: person.age,
-                           parents_permission: person.parents_permission,
-                           specialization: person.specialization, type: 'Teacher' })
-      end
-    end
-    File.write('people.json', JSON.pretty_generate(people_data))
-  end
-
-  def save_books
-    books_data = []
-    @books.each do |book|
-      books_data.push({ title: book.title, author: book.author })
-    end
-    File.write('books.json', JSON.pretty_generate(books_data))
-  end
-
-  def save_rentals
-    rentals_data = []
-    @rentals.each do |rental|
-      rentals_data.push({ date: rental.date, book: rental.book.title, person: rental.person.name })
-    end
-    File.write('rentals.json', JSON.pretty_generate(rentals_data))
   end
 end
